Move faucets declarations to core-integration for sharing
